+++ src/src/helpers/story-helpers/infoBlocks/StoryInfoBlock.jsx
@@ -1,0 +1,11 @@
+import AnyFieldUrlSelectionStoryInfoBlock from "./AnyFieldUrlSelectionStoryInfoBlock";
+
+const infoBlocks = {
+    AnyFieldUrlSelection: AnyFieldUrlSelectionStoryInfoBlock
+};
+
+export function StoryInfoBlock({infoBlockName, options}) {
+    if(infoBlockName) { 
+        return infoBlocks[infoBlockName](options);
+    }   
+}