+++ src/src/stories/anyFieldTypeAheadStories.stories.jsx
@@ -2,7 +2,8 @@ import {
   generateAnyFieldStoryDefaultExport,
   AnyFieldStoryTemplate as Template,
   standardSelectionAnyFieldArgs,
-  standardAnyFieldSelectionArgTypeConfiguration
+  standardAnyFieldSelectionArgTypeConfiguration,
+  urlSelectionAnyFieldArgs
 } from '../helpers/story-helpers/anyFieldStoryHelpers';
 
 
@@ -21,7 +22,6 @@ export default generateAnyFieldStoryDefaultExport(anyFieldStoryDefaultExportOpti
 
 
 const standardTypeAheadAnyFieldArgs = {
-  ...standardSelectionAnyFieldArgs,
   multiple: false,
   checkbox: false,
 };
@@ -29,14 +29,33 @@ const standardTypeAheadAnyFieldArgs = {
 export const TypeAheadChoiceField = Template.bind({});
 export const TypeAheadObjectField = Template.bind({});
 
+export const TypeAheadUrlChoiceField = Template.bind({});
+export const TypeAheadUrlObjectField = Template.bind({});
+
 TypeAheadChoiceField.args = {
   label: "Type Ahead Choice Field",
+  ...standardSelectionAnyFieldArgs,
   ...standardTypeAheadAnyFieldArgs,
   type: 7,
 };
 
 TypeAheadObjectField.args = {
   label: "Type Ahead Object Field",
+  ...standardSelectionAnyFieldArgs,
+  ...standardTypeAheadAnyFieldArgs,
+  type: 10,
+};
+
+TypeAheadUrlChoiceField.args = {
+  label: 'URL Type Ahead Choice Field',
+  ...urlSelectionAnyFieldArgs,
+  ...standardTypeAheadAnyFieldArgs,
+  type: 7
+};
+
+TypeAheadUrlObjectField.args = {
+  label: 'URL Type Ahead Object Field',
+  ...urlSelectionAnyFieldArgs,
   ...standardTypeAheadAnyFieldArgs,
   type: 10,
 };